move spectate page to pages
