+++ src/renderer/pages/spectate/footer.tsx
