+++ src/renderer/pages/spectate/spectate_item.tsx
