+++ src/renderer/pages/spectate/spectator_id_block.tsx
